refactor(cli): split orchestrator run runtime modules
Summary:
- replace the monolithic orchestrator_run.rs with a directory module
  rooted at orchestrator_run::run_command
- extract runtime config loading into config.rs, tracker/workspace/
  worker adapters into backends.rs, and snapshot mapping into
  snapshot.rs
- keep the existing unit and integration coverage in place while adding
  module ownership comments for the new boundaries

Rationale:
- reduce the maintenance risk called out in review without changing the
  opensymphony run behavior or public CLI surface
- make the CLI runtime seams explicit so future changes to config,
  backend adapters, and control-plane mapping can evolve separately

Tests:
- cargo test -p opensymphony-cli orchestrator_run
- cargo test -p opensymphony-cli --test run
- cargo fmt --check
- cargo clippy -p opensymphony-cli --all-targets -- -D warnings
